Add build-patch target
This new Phing target allows to build distributable ZIP archives for
patch updates, such as 1.6→1.6.8 and 1.6.7→1.6.8.

The target expects two properties, namely version and patchee. The
former is the same as for the build target, the latter is supposed to
hold the filename of the full CMSimple_XH distribution the patch should
be build against at. Both the patchee and the current release ZIP files
are supposed to exist.

Obviously, this target could be improved in several directions, but we
leave that for another time. For now we consider it sufficient, that it
works.


git-svn-id: http://svn.code.sf.net/p/cmsimplexh/code/branches/1.6@1690 4f8f9682-67af-44ce-aacd-401fd16d96b5