Fix broken info link to cmsimplewiki.com
cmsimplewiki.com is dead, so linking to it from the sysinfo doesn't make
sense. Linking to the archived cmsimplewiki.com appears to useless, as
it is out-of-date and won't be updated anymore. Instead of removing the
link, we replace it with a link to the *German* "Alles auf einen Blick"
overview page.


git-svn-id: http://svn.code.sf.net/p/cmsimplexh/code/branches/1.6@1694 4f8f9682-67af-44ce-aacd-401fd16d96b5