Fix #7: Don't show generator=CMSimple_XH in HTML header
For security reasons we don't want to show the CMSimple_XH version and
info about installed plugins in the generated HTML sources by default.
Instead we show this info only, if the debug-mode is enabled (as
suggested by olape), so supporters would still be able to view this
information without access to the back-end.