Fix #3: Search doesn't cater to alternative headings
page_params replaces the alternative heading only for the current page,
so the search functionality completely ignores the alternative headings
except for the first page. Replacing all page headings for each request
would be wasteful, so we factor out
Pageparams_replaceAlternativeHeading(), and use it from XH_Search, if
it's available. This introduces a soft core dependency on page_params,
but we prefer that over adding a respective function to the core which
would only be useful for page_params, and we keep in mind that we will
want to integrate at least parts of page_params into the core, anyway.