Fix #17: filebrowser: attempt to delete a unempty directory
Since we type-hinted array parameters (commit 4f0206fc), we have to
make sure that we pass an array as second argument to
Filebrowser\View::error(), ::success() and ::info(), but never a
string.

We also type-hint Filebrowser\View::translate()'s second parameter,
what we should have done in the first place as most calls to
::translate() happen through the other methods, and we can now nicely
simplify this method.