Close #21: Don't import arbitrary globals from the query string
We only import white-listed globals from the query string for security
and sanity reasons. To keep the BC break as small as possible, we add
the globals used by the core to the white-list, and stick with the
current way to trigger the plugin administration. Plugins already using
`XH_wantsPluginAdministration()` are supposed to work as before. Other
extensions will have to be adapted, especially if they're making use of
this global import "feature" elsewhere (i.e. not only for the plugin
administration).

Particularly note that the TinyMCE plugin's administration is broken by
this commit, but this could easily be fixed, and TinyMCE is most likely
going to replaced anyway.