Close #116: SystemCheck to state present state
We don't actually change the system check to state the present state,
but rather rework it to make it clear that we're checking for a desired
or required condition, and whether that check succeeded, failed or
raised a warning. This way we can also get rid of the explicit <img>
elements, and instead re-use the respective core classes. We also
refactor a bit.