Close #132: Move editorhooks to respective plugin
We move the editorhooks of the default filebrowser from its plugin
directory into the directory of the respective editor plugin. This
does neither require to ship CMSimple_XH with editorhooks for known
editors, nor that unknown editors have to install a file in a foreign
plugin folder.

For maximum backward compatibility we keep the existing editorhooks of
third-party editor plugins for now, and in case no editorhook.php is
found in the plugin directory, fall back to the old place.