Add dedicated change password page
Changing the password from the general CMS configuration form appears
to be problematic, so we're adding a dedicated "change password" page.
This doesn't solve the actual issue, but is a prerequisite to drop
support for the password input fields in the general CMS configuration.

We add ChangePassword, which implements the core responsibilities, and
is basically a Model/View/Controller in a single class. Of course, this
screams for refactoring, as does most of the already existing rest of
the code base.