Completely rid the `password` config type
We have already resolved the issues related to security→password, but
`password` config options generally would have these issue, so we
completely rid the `password` type. This is a BC break, but supposedly
no extension is using this type anyway, so it seems to be okay for a
minor version.

This also closes #145: No synchronous XHR for password verification.