Close #141: place template stylesheet at the end
It makes sense to link the template stylesheet after the plugin
stylesheets, so that the template has the last word (without refraining
to !important or specificity trickery). This also would be helpful with
regard to #140, because at least combining core.css and plugins.css
could easily be done without any issues regarding language or even page
specific templates.

Technically, this constitutes a BC break, but we don't think there are
too many issue in practice.