Close #130: Add length indicator to site title input
A page title shouldn't be too long; some recommend at most 55
characters. Therefore we add a length indicator to the meta_tags title
input element, and explain that in the respective language string.

We do not add a length indicator to `$tx[site][title]`, because the
actual page title depends on the `$cf[title][format]`, and can't
therefore be calculated generally. Furthermore, restricting the title
length is primarily a SEO issue, and for this purpose, setting
indivdual titles via meta_tags appears to be appropriate, anyway.