Close #133: Add check for "secure" password
We compute and display the score of passwords immediately while they
are entered in the change password screen, because that appears to make
the most sense, as users can improve the password right away, instead
of checking the password score in the system check and changing the
password again if necessary. We also compute and display the password
score of the old password for reference (particularly interesting wrt.
default password).

We display the password score in a simple textual form instead of a
fancy password strength meter for accessibility and simplicity.

The algorithm to calculate the password score was taken from
http://stackoverflow.com/questions/948172/password-strength-meter#11268104
with minor improvements. It is very simplistic, but appears to yield
sensible results, e.g. scores > 100 indicate very strong passwords. The
algorithm works properly for ASCII characters only, but our passwords
are limited to a subset of ASCII anyway, so that's fine.