Close #125: Internationalize date representations
If class `IntlDateFormatter` is available, we use it to format date/times
in `XH_formatDate()`; otherwise we fall back to the old solution using
`$tx['lastupdate']['format']`.

The advantage of `IntlDateFormatter` over `strftime()` is that ICU has
its own locale support, so it's not necessary to have the locales
actually installed, and `setlocale()` has issues in multi-threaded
environments anyway.