Close #77: Partially integrate page_params into the core
At least the basic functionality regarding publishing and hiding of
pages (i.e. the page_params alternatives to `#cmsimple hide#` and
`#cmsimple remove#`) shouldn't be handled by a plugin, but rather by
the core, because:

* `#cmsimple remove#` can't be detected by a plugin as it is already
  converted during content loading
* during plugin loading `hide()` may yield wrong results
* other plugins should have a simple means to detect whether a page is
  hidden or unpublished

Therefore, we extract the functionality to detect whether a page is
hidden and published into the new class Publisher whose sole instance
is globally available as `$xh_publisher`. As Publisher is perfectly
capable to determine the first published page, we don't have to track
that in the global `$_XH_firstPublishedPage` anymore, so we remove this
variable, what is a potential BC break, but we're convinced that the
benefits of encapsulating the hide/unpublish functionality in a class
clearly outweighs the drawbacks of this BC break.

Future scope: Publisher may also take the responsibility to modify the
page content for unpublished and hidden pages, and might offer an API
for third-party plugins to mark pages as hidden or unpublished.