Close #67: Page_params: add date-time picker
Instead of adding a fancy JavaScript date/time picker, we use native
datetime-local input elements which fall back to plain text inputs if
not supported, in which case we do the validation manually, as before.

Support for native datetime-local input elements is not so bad
currently, and Firefox 54 is supposed to add support if enabled in the
configuration.

Anyhow, using datetime-local input elements slightly changes the
required input format and makes it also more strict (e.g. the time is
mandatory), so we introduce a BC break which has to be dealt with by
Exchange_XH.