Fix #200: CMS browser throws error when opening a file
This regression has been introduced when removing the event handler
attributes with commit bef4789. Of course, we must not attach
onFileChoose() as click event listener for the CMS browser, but only
for the editor browser. In lack of a better way to distinguish both
browser types, we check for #filebrowser_subfolders which exists only
in the CMS browser.