Close #68: Replace bundled templates
We've decided that CMSimple_XH's default distro should have a fresh
look, so we're replacing the 1.6 templates with three new ones. Thanks
to frase, lck and richukunst for providing the new templates.

As the new distro content is already work in progress, we don't care to
adapt it yet with regard to the subpages of "Templates and Plugins".