Fix: filebrowser deletes files linked in unpublished pages
The filebrowser isn't able to detect files used on unpublished pages,
because it is running in normal mode. We change that to edit mode, what
appears to work fine with the default filebrowser, but we're not sure
whether external filebrowsers will have issues with that. Anyway, we'll
have a beta, so we try it out.

See #207.