Close #92: Improve accessibility of the admin menu
The drop-down menus of the admin menu are not accessible on touch
devices and via the keyboard. We therefore make all functionality also
accessible from the toplevel items of the admin menu, i.e. we add yet
missing links to the settings page and introduce a new plugins page
with links to all enabled plugins and their registered menu items.