Close #210: Improve languagemenu flag icons
The current language flags are tiny (18*12px), and so the language menu
is practically unusable on mobile devices. We therefore replace them
with larger icons (32*24px). Furthermore, we allow templates to ship
their own flag icons in the flags/ subfolder of the template, and fall
back to userfiles/images/flags, if a flag icon isn't available in the
template.