Close #140: Combine all stylesheets
As there are issues regarding the template stylesheets
(import-at-rules and/or multiple templates) we exclude them, but at
least we combine `core.css` and `plugins.css` in the file
`stylesheet.css`. This way we save one request, what's still a nice
improvement.

Instead of fixing the related unit test, we dump it because it relies
on the deprecated ::assertTag().