Close #204: Password hashing, again
We replace the out-dated PHPass with PHP's `password_*` and
`random_bytes` functions, and provide a slightly modified version of
<https://github.com/ircmaxell/password_compat> as fallback for PHP
versions which do not natively implement said functions.

For best portability we use `PASSWORD_BCRYPT` as $algo, which is equal
to `PASSWORD_DEFAULT` for PHP 7.1 at the very least,  anyway.

The `random_bytes` fallback is weak for some systems, but still better
than the routine provided by PHPass.

We also rehash the admin password, to make use of the improved hashing.