Call variable function directly
There's no need for `call_user_func()`, and calling the variable
function directly is even a bit faster.

The main reason for this change, however, is that uopz currently
doesn't trigger hooks that have been set, if the function is called via
`call_user_func()`, and we're planning to use it to replace runkit for
PHP7.