Fix #265: cmsimple/classes/Menu.php has been removed
There are several places where it has been documented to include this
file manually to use `XH_Li`, and since we had renamed the file in
e4807ea, we had introduced an unintended BC break. Therefore we add the
file back, so that including it doesn't trigger a fatal error, but
instead a deprecation notice.

Actually loading `Li.php` as has been suggested, is not necessary, as
this is catered to by the class autoloader.