Put XH_debug() under test
We are trying to test everything except the failing guard at the start
of the function – actually, this guard appears to be testing PHP,
because the error handler is not supposed tp be called if
`error_reporting` does not contain the current error type.

We change `die()` to `XH_exit()` to be able to mock the function.