Do not explicitly document @globals and @staticvars of functions
The benefit of explicitly documenting which global variables are used
by a certain function (or method for that matter) is rather doubtful,
since we do not document other entities (functions, classes, constants)
which are used in a function either.  Furthermore, properly maintaining
these docs takes some time, and is prone to error nonetheless.  Also
these information could at least theoretically be automatically gained
from a static analysis tool.

It might make sense to document global variables which are modified by
a function, but we are not aware of any documentation tool actually
supporting `@global`.  At least neither Doxygen nor phpDocumentor do,
so we would have to work around this limitation to avoid broken docs.
This time is better spend to get rid of the globals altogether, or at
least encapsulate each in a single spot.

The situation regarding static function variables is somewhat similar.
`@staticvar` is not supported by the relevant documentation generators,
and it is a doubtful feature anyway.  It would be better to refactor
such functions to methods, and to turn the static function variables
into static class variables.

Therefore we rid all `@global` and `@staticvar` doc block comments
regarding functions and methods.  Of course, we keep the `@global`
doc block comments of the actual variables, of course.