Remove @category and @package doc block tags
These legacy markers support(ed) (pseudo) namespaces.  However, since
proper hierarchical namespaces are supported as of PHP 5.3 they are
superfluous now at best.  `@category` has already been deprecated in
phpDocumentor, and means something completely else in Doxygen.
`@package` can still be used, but phpDocumentor recommends against
using this tag, if logical and function subdivision is equal, and
Doxygen even thinks that entities in files with this tag belong to that
namespace, what is not the case for our functions.