Strengthen Phan analysis and fix reported issues
Mostly, these are minor typing related issues, but we also caught some
bad practises (such as not declaring abstract methods in a base class),
found some occurrences of tag() which have been missed, and even found
an actual (albeit minor) bug (see #307), where an unusable content
folder may be created if it doesn't already exist.  We are going to fix
that latter issue with a separate commit since it has to be applied
against 1.7 as well.

Besides the mentioned bug, a single reported issue remains, which
appears to be a bug in Phan, and needs closer investigation.