Close #316: Support template specific configuration and i18n
We introduce the global variables `$tpl_cf` and `$tpl_tx` which work
basically like `$cf` and `$tx`, i.e. they are loaded from the active's
template `config/` and `config/languages/`, respectively, if available
(otherwise the variables hold empty arrays).  `defaultconfig.php` and
`default.php`, respectively, are also supported.  Both variables are
supposed to hold "two-dimensional" arrays (mostly for conformance with
`$cf` and `$tx`) where the first dimension is an arbitrary category
name.  The `help` category of `$tpl_tx` is special, and allows to add
help texts for the respective `$tpl_cf` variables.  `metaconfig.php` is
also supported which allows for typed configuration options.

`$tpl_cf` and `$tpl_tx` can be used in the template as desired by the
designer.

We implement the possibility to edit the configuration and language file
of the current template from the back-end, and change the layout of the
settings menu accordingly.

We also introduce new `$pth` variables to abstract over the concrete
file paths of the template configuration and language files and folders.

For demonstration purposes we add the adjusted fhs-simple-config
template and make it the default template.