Fix #283: Mailform should use fixed From address
Presently, mails sent via the built-in mailform use the submitted email
address as From header.  However, unknown senders may be blocked on
shared hosting servers, and therefore we set the From header to the
configured email address (which easily can be whitelisted, if
necessary), and set the Reply-To header to the submitted email address.