Improve regexp
While the existing regexp works, it does not really implement what it
is supposed to check.  And yes, this regexp could be substituted by a
`stripos()` check, but firstly, the regexp may be fine-tuned in the
future, and secondly, the cmsimple/ folder is supposed to be protected
from direct HTTP access anyway.