Fix #343: Fatal errors: nonsensical handling
If an `E_PARSE` or `E_ERROR` error occurs, and debug mode is enabled,
we emit basically the same error message as PHP would have done, if
`display_errors` had been enabled.

We prefer this potential information leakage over letting the user
guess what has gone wrong, since debug mode is not supposed to be
enabled all the time in production environments, and since such kinds
of errors usually only happen after installing a new component or after
switching the PHP version, so the first to notice the problem is
hopefully the webmaster, anyway.  It also should be noticed, that
indeed only fatal errors will be shown this way, but not notices and
warnings, etc., what further limits the information leakage.