Remove superfluous code
If `$i == 0 && $x == ''` the returned link is identical to what would
be returned without the special casing, since $u[0] is guaranteed to
exist.  Even if the latter was not given, the link is not depending on
the admin modus or the locator → show_homepage setting – the old code
simply makes no sense.