Fix #352: Filebrowser: NOTICE: getimagesize(): Read error!
We revert commit c5ae03147f5735fc07fcaadc86d6b586087dd248, because it
was simply wrong.  We have to check whether the image files are
supported by `getimagesize()`, regardless of whether they are
configured as allowed image extensions.

In the long run we may want to support image previews of image types
which are not (yet) supported by `getimagesize()` (for instance, SVG,
see #294), but for now we stick with the formerly working code.