Fix #366: Search highlighting may break markup
HTML manipulation with regular expressions is too error prone, as the
bug report shows; manipulating the DOM is way more resilient.  Since
the highlighting of the search strings is not a fundamental property of
CMSimple_XH, we change it to be done as progressive enhancement.

We're using a JavaScript routine which is based on Stuart Langridge's
serchhi.js[1], but use only those parts which we really need.  We also
refactor these parts according to our needs.  Since we still support
IE8 for the back-end, we stick with this for the front-end search
highlighting.  We also do not want to move functionality from
admin.min.js to core.min.js in a patch version, so we write even more
clumsy code which we should refactor for XH 1.8.

[1] <https://kryogenix.org/code/browser/searchhi/>