Drop <phpunit> task in favor of direct execution
The direct phpunit execution output is more sensible than what phing
makes of it; there is no need to see the names of the tests that have
been executed, and the test matrix displayed by phpunit gives a great
summary.

Furthermore, it is especially nice that the error locations are now
better understood by IDEs, so that jumping to the code is just a click
away.