Fix #315: Remove CMS/Site Info
In CMSimple_XH 1.7.0 the implicit CMS/Site Info page has been
introduced, but it appears this has not been a good idea regarding
security, because the page offers to much information regarding the
CMS, and installed templates and plugins, which might be exploitable.

Therefore we remove this page completely, but keep poweredByLink() for
backward compatibility, changing it to simply emit a link to
cmsimple-xh.org.