Adjust build target for reqcheck.php
The distribution packages should have reqcheck.php outside of
cmsimplexh/, because the script is not actually part of the CMS.  We
also use a placeholder for the CMSimple_XH version, and fill that in
during build time.

While we're at it, we also remove the extraneous replacement of the
placeholder for the old tinymce plugin, which is superseeded by
tinymce4 as of CMSimple_XH 1.7.0.