Fix conceptionally flawed ::assertXPathContains()
This assertion is supposed to verify that some `$text` is contained in
any of the nodeValues of a set of nodes.  However, it has been written
in a way, that it always passed for empty `$text`s, which defeated its
purpose.  In particular, `AdminMenuTest::testShowsItem()` always
succeeded, because data providers are always run before `::setUp()`,
so that the `$tx` mocks were always empty.

We fix this by requiring the English language file for AdminMenuTest.
We also fix some minor missing setup.