Fix #484: page_params: radio buttons may not be preselected with PHP 8
The values we receive are actually strings, and an empty string is
supposed to be equivalent to "0".  However, PHP 8.0 actually does a
string comparison in this case, so we cast to int, to get the desired
results with all PHP versions.

While we're at it, we also apply stricter comparisons regarding that
setting in the front-end, although the behavior doesn't seem broken.