Fix #480: CSS of updated plugins may be ignored
CMSimple_XH generates a combined stylesheet of the core and all plugin
stylesheets.  This cached version expires whenever plugins are added or
removed, or when any plugin stylesheet is newer then xhstyles.css based
on the filemtime.  If a plugin is updated, its stylesheet's filemtime
may be older than that of xhstyles.css, so the actually modified file
may just be ignored.

We solve that by comparing SHA1 hashes instead, what is less efficient,
but should avoid the current cache eviction issues.