Improve support for soft hyphens in page headings
While it is already possible to include soft hyphens in page headings
by manually inserting the SOFT HYPHEN (U+00AD) character, these are
usually not displayed when editing page headings, although soft hyphens
appear to be particularly useful for page headings.

Thus, we allow to configure an arbitrary string to stand in for a soft
hyphen, which is, however, omitted from the page URL.  If that config
option is empty (the default), everything stays as it was before.