Revive Phan
This has apparently been neglected for quite some time; new version
show a bunch of issues, which we're not comfortable to fix yet.  Thus
we take a step back and only report critical issues.

We also add the `phan` check to CI, so this does not happen again.  We
use Phive to install Phan, to avoid the potential composer require-dev
dependency hell.