Fix case-sensitive parsing of Location headers
HTTP headers are supposed to be case-insensitive, so `location:` is
perfectly valid; we need to cater to that.

See also <https://cmsimpleforum.com/viewtopic.php?t=17678&p=80666#p80667>.

This patch had been provided to hi_updatecheck as PR[1], but since
there has been no feedback so far, and this bug affects quite some
users, we apply it to the bundled hi_updatecheck.

[1] <https://github.com/TN03/hi_updatecheck/pull/5>