Remove empty selected from query string
Typically, GET forms that should be submitted to the same page use a
`<input type="hidden" name="selected" value="$su">`.  Such GET requests
(`?selected=page`) are then permanently redirected to the respective
page.  However, that fails on the first page where `$su` is empty, and
results in `?selected=`.

Instead, we need to redirect even if `$selected` is empty, and to make
sure that we remove the query parameter in that case, and also not to
have a trailing `?`.