Fix #546 (TinyMCE 5 is not shown on the product pages of XHShop)
Plugins may call `editor_replace()` passing a configuration file or
shortcut as `$config`.  The respective editor implementations need to
figure that out, and convert that to an appropriate initialization
script.

We also make sure that we don't provide a broken initialization script,
in case the requested init file does not exist.  This prevents
JavaScript errors from being logged.