Upgrade to PHPUnit 8
Versions prior to that are not compatible with PHP 8, but we need to
be able to develop with this PHP version.  The downside of the upgrade
is that we now require PHP ≥ 7.2.0 for development.

We also need to add a general PHPUnit configuration, not only for code
coverage, and do some minor adaptions for the new PHPUnit version.