Plugincalls returning null|void emit deprecation notices
The return value of plugin calls is passed to `preg_replace_callback()`
unmodified, and automatically coerced to a string, because that is what
`$subject` expects.  This works fine for all scalar values, and also
for `null` (or even if the function does not return a value at all)
prior to PHP 8.1.0.  However, as of that PHP version, coercing
`null|void` to `string` is deprecated.  We fix that by casting to
`string` explicitly.