Option not to publish changes immediately 1.8
https://github.com/cmsimple-xh/cmsimple-xh/issues/609

A backup copy to be published can be entered in the language file here.
Settings -> Language -> Publish -> Current
i.e.: 20240724_105045_extra.htm
It makes sense to create a manual backup so that this copy is not deleted by XH routines.
Now you can make unhindered changes to the content without it being permanently published.
The current content is only published when the entry in the language file is removed or the manual backup is deleted.