Add some kind of undo feature (#608)
* Add some kind of undo feature

https://github.com/cmsimple-xh/cmsimple-xh/issues/51

An additional temporary backup is created for each saving process that affects the content. These backups can be handled in the same way as all other backups. All temporary backups are deleted when you log out.

The option can be configured under "Settings - Backup - Temporary backup".

* Option not to publish changes immediately 1.8

https://github.com/cmsimple-xh/cmsimple-xh/issues/609

A backup copy to be published can be entered in the language file here.
Settings -> Language -> Publish -> Current
i.e.: 20240724_105045_extra.htm
It makes sense to create a manual backup so that this copy is not deleted by XH routines.
Now you can make unhindered changes to the content without it being permanently published.
The current content is only published when the entry in the language file is removed or the manual backup is deleted.

* Manually created backup copies can be deleted directly

Previously, these backups could only be deleted via FTP or were deleted by a restore operation.

* Option not to publish changes immediately with dropdownfield 1.8

https://github.com/olape-git/cmsimple-xh/commit/2585f2bc99c1f161854c3362c6c5e695ae7f4e8a

The corresponding file to be published can now be selected via the drop-down field.