Warning when checking for access protection
In certain cases it can happen that a warning is issued when checking for access protection via cURL, even though the file is protected. This happens, for example, if your own error pages are set up via .htaccess. Then the status code is not between 400 and 500, but 302.
I therefore suggest extending the cURL options so that the first redirect is followed.

https://github.com/cmsimple-xh/cmsimple-xh/issues/607