printlink for newsbox (#628)
It is therefore possible to include a second parameter when calling the newsbox() function. Instead of <?=newsbox('News01');?> then <?=newsbox('News01', 'printlink');?>. This will display the link to the print preview in the newsbox (only for the content of this newsbox).

It is also possible to add a url to the printlink() function. For example, printlink('Test/Test page'). This could be useful for content in Expandcontract_XH, for example.