Drop random_bytes() fallback (#647)
CMSimple_XH 1.8 requires PHP 7.4, so there is no need anymore to
provide a fallback for `random_bytes()` which is available as of PHP
7.0 anyway.

And even if CMSimple_XH 1.8 was still compatible with PHP 5.6, on
Windows the fallback would yield very bad results, unless OpenSSL was
available.  Better to error out, than yielding random values which
don't hold up to their promise.