Move executable statement out of adminfuncs.php
The whole point of adminfuncs.php is to contain functions only (and
maybe other declarations), so that it can be included by clients
without actually executing any code.

Thus we move the statement to cms.php right after adminfuncs.php has
been included there, having the same effect as if it was contained in
adminfuncs.php.