Fix BackupTest
The optional parameter `$regularOnly` of `XH_isContentBackup()` has
been changed from bool to string[1], so we need to update the test
accordingly.

While we're at it, we also update the doc block to mention the new type
but we stick with the misleading parameter name, to avoid a BC break in
a patch version.

[1] <https://github.com/cmsimple-xh/cmsimple-xh/commit/7975ff20c80fb2ee09aa739743cebff5c203ab1f>