Rehash password if necessary (#656)
* Rehash password if necessary

We're using the default parameters of `PASSWORD_BCRYPT`, and especially
the `cost` parameter increased for more recent PHP versions what is a
consequence of ever increasing computing power.  However, if a password
has been hashed with an older PHP version, or a user reuses a hash that
has been created by an old CMSimple_XH version, the `cost` factor might
be as low as `10` or even only `8`, while PHP 8.4 uses `12` by default.

To cater to this, we check whether the password needs to be re-hashed
on successful login with a non-default password, and if so, re-hash the
password.  To be able to actually store the new hash in config.php, we
move the respective method from `ChangePassword` to the `Controller` so
we can re-use the non trivial logic.

* Avoid warning `Undefined array key "xh_default_password"`

When we save the configuration, we need to set the respective session
variable.

An alternative would be to check whether the session variable is set
before accessing it in `XH_writeFile()`.

* Use fix cost of 12 for password hashing

Otherwise password will be re-hashed for older PHP versions (i.e. prior
to PHP 8.4.0), thus weakening brute-force resilience.

* Restore config.php