Defer $hjs injection till final clean-up
Traditionally, `$hjs` has been written to `<head>` when `head()` was
called.  That makes `$hjs` basically unusable if a plugin call is
supposed to be put in the `<body>` of the template.  While there are
some workarounds, sometimes elements must be placed in `<head>`.

Therefore, we do no longer write `$hjs` directly in `head()`, but only
put a placeholder there, which we replace in `XH_finalCleanUp()`.

Alternatives:

* don't use a placeholder, but replace like it's done for `$bjs`
* use a "simpler" placeholder than an UUID, e.g. `<!--$hjs here-->`

We're choosing to do it this way, to (a) provide maximum backward
compatibility (changing the order of elements in `<head>` might cause
issues), and (b) to further reduce the risk that the placeholder is
inadvertently overwritten by a plugin (highly unlikely, though).

See also <https://cmsimpleforum.com/viewtopic.php?t=13988&p=92323#p67412>.