Test fixes (#659)
* Fix LinkCheckerTest

The respective config options have been renamed[1], but the test case
has not been updated so far.

[1] <https://github.com/cmsimple-xh/cmsimple-xh/commit/107d4a4d8ff82c0558206fb2654389960ac1b62e>

* Fix BackupTest

The optional parameter `$regularOnly` of `XH_isContentBackup()` has
been changed from bool to string[1], so we need to update the test
accordingly.

While we're at it, we also update the doc block to mention the new type
but we stick with the misleading parameter name, to avoid a BC break in
a patch version.

[1] <https://github.com/cmsimple-xh/cmsimple-xh/commit/7975ff20c80fb2ee09aa739743cebff5c203ab1f>

* Fix HeadTest

The test had been written before $cf[head][links] has been
introduced[1], so we need to set up the config option now.

[1] <https://github.com/cmsimple-xh/cmsimple-xh/commit/94fb7e42d1107fa6d29a2b771fdbed72edd6433d>

* Fix ControllerLoginTest

Due to changes to the login[1] we need to adapt the test fixture.

[1] <https://github.com/cmsimple-xh/cmsimple-xh/commit/2cdf2572f578ecf5f3170320acda93c9576a15e3>

* Fix PluginTextFileEditTest

Due to changes to `XH_writeFile()`[1][2], we need to add to the test
fixture.

[1] <https://github.com/cmsimple-xh/cmsimple-xh/commit/22c4bfce418a3c28c6fb819cac4b1a46b9966d47>
[2] <https://github.com/cmsimple-xh/cmsimple-xh/commit/6a873583679a6f9272bedbcac62108345689c58a>

* Fix PageDataRouterTest and PageDataModelTest

Due to changes to to `uenc()`[1], we need to amend the test fixture.

[1] <https://github.com/cmsimple-xh/cmsimple-xh/commit/c8ad5b797857dd8ed09f6a83dd033c576ced1f1c>

* Fix ErrorHandlerTest

Due to changes to `XH_debug()`[1] we need to amend the test fixture.

[1] <https://github.com/cmsimple-xh/cmsimple-xh/commit/500099e97f5a686790c1daba6e9194b09192f6c7>

* Fix EmergencyTemplateTest

`XH_emergencyTemplate()` expects a string argument now[1], so we pass
an empty string.  We document that parameter right away.

And we also need to mock `languagemenu()` which is now called by the
emergency template.

[1] <https://github.com/cmsimple-xh/cmsimple-xh/commit/90778d27ab90c4101da1ac0053c6de83a59a15e9>

* Fix remaining tests

Due to changes to `XH_writeFile()`[1][2], these require an amendment to
the test fixture.

[1] <https://github.com/cmsimple-xh/cmsimple-xh/commit/22c4bfce418a3c28c6fb819cac4b1a46b9966d47>
[2] <https://github.com/cmsimple-xh/cmsimple-xh/commit/6a873583679a6f9272bedbcac62108345689c58a>

* Actually enable the tests in CI

Also need to run on Ubuntu 22.04, since Ubuntu 20.04 is in brownout.