Upgrade some dev tools
First, we upgrade to Phing 3, what requires PHP ≥ 7,4 which is not a
problem since that version is required by CMSimple_XH 1.8 anyway.

Then we need to upgrade to PHPUnit 9, since otherwise there are install
conflicts.  That requires us two touch two test cases, because we need
to reset a function scope static variable.

We now also run CI (if it runs) on PHP 8.2; PHP 8.3 and 8.4 are likely
out of reach for now, since there are issues regarding uopz.