Introduce SystemInfo class
We move the implementations of `XH_sysinfo()`, `XH_systemCheck()`, and
`XH_systemCheckLi()` into the new class `SystemInfo`.  To avoid a BC
break, we keep the class stateless, and make the methods public; this
way we can just call them from the functions (which are now thin
wrappers).

We do not bother with the heavy use of global variables, but extract a
couple of thin wrappers over functions which are inherently untestable.
We can now use a test-specific subclass[1] to bring most of the class
under test.

Since the whole point of the class is to render the system info screen,
we employ an approval test[2] which verifies the output exactly, but
can easily updated if the expected output changes.

Note that we neither subclass the test class from our own `TestClass`
nor that we use uopz (or other dark magic) there.  Also note that we
now have a test that verifies the default password, so that it can't
inadvertently be changed without noticing.

[1] <http://xunitpatterns.com/Test-Specific%20Subclass.html>
[2] <https://approvaltests.com/>