Fix password forgotten feature
This was broken by 3f6f770f27c4d3482d8e1205f12705915f81f280.

Obviously, when resetting the password, there is no session, so PHP 8
will throw a `TypeError` since `$_SESSION` is `null`.  The fix is
trivial, but we take the opportunity to bring that functionality under
test.  To avoid sending an actual mail, we allow to inject a `Mail`
object in the constructor (if none is passed, everything behaves like
before), and stub the `::send()` method.  We then verify that the
password has actually been changed in config.php.