Fix #654: Redirection due to selected is broken as of PHP 8.0.0 (#657)
We make sure that `XH_ADM` is actually defined before we access it to
avoid a fatal error in PHP 8, in case we exit early, before `XH_ADM` is
actually defined.

We do not unset the password hash stored in the session in this case,
since that would lead to unintentional logout.